fix warnings

+++ Geo/GModel.h
@@ -397,7 +397,7 @@ public:
   piter lastElementaryName() { return _elementaryNames.end(); }
 
   // get the number of physical names
-  int numPhysicalNames() const { return _physicalNames.size(); }
+  int numPhysicalNames() const { return (int)_physicalNames.size(); }
 
   // get iterators to the last physical name of each dimension
   void getInnerPhysicalNamesIterators(std::vector<piter> &iterators);
@@ -680,9 +680,9 @@ public:
   int writeSTL(const std::string &name, bool binary = false,
                bool saveAll = false, double scalingFactor = 1.0,
                int oneSolidPerSurface = 0);
-  
+
   // X3D (only output from OCCT's triangulation)
-  int writeX3D(const std::string &name, 
+  int writeX3D(const std::string &name,
                bool saveAll = false, double scalingFactor = 1.0,
                int oneSolidPerSurface = 0);
